test: merge command distributor tests
