+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/distribution/CommandRedistributorTest.java
@@ -7,68 +7,177 @@
  */
 package io.camunda.zeebe.engine.processing.distribution;
 
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import io.camunda.zeebe.engine.state.immutable.DistributionState.PendingDistributionVisitor;
+import io.camunda.zeebe.engine.metrics.DistributionMetrics;
+import io.camunda.zeebe.engine.processing.streamprocessor.writers.Writers;
+import io.camunda.zeebe.engine.state.appliers.EventAppliers;
+import io.camunda.zeebe.engine.state.mutable.MutableDistributionState;
+import io.camunda.zeebe.engine.state.mutable.MutableProcessingState;
+import io.camunda.zeebe.engine.state.mutable.MutableRoutingState;
 import io.camunda.zeebe.engine.state.routing.RoutingInfo;
+import io.camunda.zeebe.engine.state.routing.RoutingInfo.StaticRoutingInfo;
+import io.camunda.zeebe.engine.util.ProcessingStateExtension;
+import io.camunda.zeebe.engine.util.stream.FakeProcessingResultBuilder;
+import io.camunda.zeebe.protocol.Protocol;
 import io.camunda.zeebe.protocol.impl.record.value.distribution.CommandDistributionRecord;
-import org.junit.Before;
-import org.junit.Test;
+import io.camunda.zeebe.protocol.impl.record.value.user.UserRecord;
+import io.camunda.zeebe.protocol.record.ValueType;
+import io.camunda.zeebe.protocol.record.intent.UserIntent;
+import io.camunda.zeebe.stream.api.InterPartitionCommandSender;
+import java.util.Set;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 
+@ExtendWith({MockitoExtension.class, ProcessingStateExtension.class})
 public class CommandRedistributorTest {
 
-  private CommandDistributionBehavior behavior;
-  private RoutingInfo routingInfo;
+  /** Injected by {@link ProcessingStateExtension} */
+  private MutableProcessingState processingState;
 
-  @Before
+  private MutableRoutingState routingState;
+  @Mock private DistributionMetrics mockDistributionMetrics;
+  @Mock private InterPartitionCommandSender mockCommandSender;
+
+  private CommandRedistributor commandRedistributor;
+  private long distributionKey;
+  private UserRecord recordValue;
+  private MutableDistributionState distributionState;
+
+  @BeforeEach
   public void setUp() {
-    behavior = mock(CommandDistributionBehavior.class);
-    routingInfo = mock(RoutingInfo.class);
+    distributionState = processingState.getDistributionState();
+    routingState = processingState.getRoutingState();
+    routingState.initializeRoutingInfo(2);
+
+    final var fakeProcessingResultBuilder = new FakeProcessingResultBuilder<>();
+    final Writers writers =
+        new Writers(() -> fakeProcessingResultBuilder, mock(EventAppliers.class));
+
+    final RoutingInfo routingInfo =
+        RoutingInfo.dynamic(routingState, new StaticRoutingInfo(Set.of(1, 2), 2));
+
+    final CommandDistributionBehavior behavior =
+        new CommandDistributionBehavior(
+            processingState.getDistributionState(),
+            writers,
+            1,
+            routingInfo,
+            mockCommandSender,
+            mockDistributionMetrics);
+
+    commandRedistributor = new CommandRedistributor(behavior, routingInfo);
+
+    recordValue =
+        new UserRecord()
+            .setUserKey(1L)
+            .setEmail("test@test.com")
+            .setName("foo")
+            .setPassword("bar")
+            .setUsername("foobar");
+
+    final var distributionRecord =
+        new CommandDistributionRecord()
+            .setPartitionId(1)
+            .setValueType(ValueType.USER)
+            .setIntent(UserIntent.CREATE)
+            .setCommandValue(recordValue);
+
+    distributionKey = Protocol.encodePartitionId(1, 100);
+
+    distributionState.addCommandDistribution(distributionKey, distributionRecord);
+    // Add pending command distributions for partitions
+    routingState
+        .desiredPartitions()
+        .forEach(
+            partition -> distributionState.addRetriableDistribution(distributionKey, partition));
   }
 
   @Test
   public void shouldNotRetryOnFirstCycle() {
-    final var redistributor = new CommandRedistributor(behavior, routingInfo);
-    final var distributionRecord = new CommandDistributionRecord().setPartitionId(2);
+    // given
+    // when
+    commandRedistributor.runRetryCycle();
 
-    doAnswer(
-            invocation -> {
-              return invocation
-                  .<PendingDistributionVisitor>getArgument(0)
-                  .visit(123L, distributionRecord);
-            })
-        .when(behavior)
-        .foreachRetriableDistribution(any(PendingDistributionVisitor.class));
-
-    redistributor.runRetryCycle();
-
-    verify(behavior, never()).onScheduledRetry(123L, distributionRecord);
+    // then
+    verify(mockCommandSender, never())
+        .sendCommand(1, ValueType.USER, UserIntent.CREATE, distributionKey, recordValue);
+    verify(mockCommandSender, never())
+        .sendCommand(2, ValueType.USER, UserIntent.CREATE, distributionKey, recordValue);
   }
 
   @Test
   public void shouldRetryOnSecondCycle() {
-    final var redistributor = new CommandRedistributor(behavior, routingInfo);
-    final var distributionRecord = new CommandDistributionRecord().setPartitionId(2);
+    // given
+    // when
+    commandRedistributor.runRetryCycle();
+    commandRedistributor.runRetryCycle();
 
-    doAnswer(
-            invocation -> {
-              return invocation
-                  .<PendingDistributionVisitor>getArgument(0)
-                  .visit(123L, distributionRecord);
-            })
-        .when(behavior)
-        .foreachRetriableDistribution(any(PendingDistributionVisitor.class));
-
-    redistributor.runRetryCycle();
-    redistributor.runRetryCycle();
-
-    verify(behavior).onScheduledRetry(123L, distributionRecord);
+    // then
+    verify(mockCommandSender, times(1))
+        .sendCommand(1, ValueType.USER, UserIntent.CREATE, distributionKey, recordValue);
   }
 
-  private void addDistribution(final CommandDistributionRecord... records) {}
+  @Nested
+  class ScalingUpPartitions {
+
+    @BeforeEach
+    void setUp() {
+      // Simulate scaling up partition 3
+      routingState.setDesiredPartitions(Set.of(1, 2, 3));
+
+      distributionState.addRetriableDistribution(distributionKey, 3);
+    }
+
+    @Test
+    void shouldNotRedistributeToScalingPartitions() {
+      // given
+      // when
+      // run cycle twice, since first one always does not try distributing the records
+      commandRedistributor.runRetryCycle();
+      commandRedistributor.runRetryCycle();
+
+      // then
+      verify(mockCommandSender, times(1))
+          .sendCommand(1, ValueType.USER, UserIntent.CREATE, distributionKey, recordValue);
+      verify(mockCommandSender, times(1))
+          .sendCommand(2, ValueType.USER, UserIntent.CREATE, distributionKey, recordValue);
+      verify(mockCommandSender, never())
+          .sendCommand(3, ValueType.USER, UserIntent.CREATE, distributionKey, recordValue);
+    }
+
+    @Test
+    void shouldRedistributeToScaledPartition() {
+      // given
+      // when
+      // run cycle twice, since first one always does not try distributing the records
+      commandRedistributor.runRetryCycle();
+      commandRedistributor.runRetryCycle();
+
+      // then
+      verify(mockCommandSender, times(1))
+          .sendCommand(1, ValueType.USER, UserIntent.CREATE, distributionKey, recordValue);
+      verify(mockCommandSender, times(1))
+          .sendCommand(2, ValueType.USER, UserIntent.CREATE, distributionKey, recordValue);
+      verify(mockCommandSender, never())
+          .sendCommand(3, ValueType.USER, UserIntent.CREATE, distributionKey, recordValue);
+
+      // then
+      // Partition 3 is now scaled up, so it should be redistributed to
+      routingState.arriveAtDesiredState();
+
+      // run cycle twice, since first one always does not try distributing the records
+      commandRedistributor.runRetryCycle();
+      commandRedistributor.runRetryCycle();
+      verify(mockCommandSender, times(1))
+          .sendCommand(3, ValueType.USER, UserIntent.CREATE, distributionKey, recordValue);
+    }
+  }
 }
